requirements.txt
Flask==3.0.0
Flask-Dance==7.0.0
Flask-Session==0.5.0
Flask-RESTX==1.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
oauthlib==3.2.2
itsdangerous==2.1.2
